+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/PromiseRejectionTracker.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -79,4 +79,9 @@ public interface PromiseRejectionTracker {
      */
     void promiseResolvedAfterResolved(DynamicObject promise, Object value);
 
+    /**
+     * Invoked when the engine has executed all promise reaction jobs.
+     */
+    default void promiseReactionJobsProcessed() {
+    }
 }